Allow reshaping an inhomogeneous system to a single unit cell (#191)
Before this commit, reshaping an inhomogeneous system was disallowed. Now, it is allowed in certain specialized contexts. The primary motivation is enable LSWT-KPM for large, inhomogeneous systems. To support this, we need to allow reshaping of an inhomogeneous system into a single, enlarged "crystal" unit cell. With the refactors, it also becomes possible to call `repeat_periodically` on an inhomogeneous system. As a design decision, `reshape_supercell` is still disallowed for inhomogeneous systems, because there is not yet an obvious use case for this.

Co-authored-by: David A. Dahlbom <34151623+ddahlbom@users.noreply.github.com>