Print an error if non-positive dt is detected (#192)
Print an error if non-positive `dt` is detected when constructing an integrator. A negative timesteps can be attained with a mutable update, e.g. `dyn.Δt *= -1`.

Also deprecate `Base.copy(::Langevin)`. We might add it back depending on the resolution to https://github.com/SunnySuite/Sunny.jl/issues/193. The story here will continue to evolve, but for now it's safe to err on the side of a thinner public API.