Track spacegroup setting (#325)
Track the affine transformation that defines the spacegroup setting. This allows look-up of Wyckoff data for a given atom position.

Crystals constructed from a spacegroup number (1 .. 230) will now be interpreted in the ITA standard setting. This makes it easy to cross-reference with the Bilbao crystallographic server. To select a non-conventional ITA setting, a `choice` argument can be passed to the `Crystal` constructor.