Updates for Makie v0.24 (#393)
Replace `arrows` with the `arrows3d` function introduced in Makie v0.23. Bump the minimum required Makie version to v0.24 to get additional bugfixes.

When plotting spins of vanishing magnitude, the arrow geometry now shrinks to zero. This is important when visualizing systems in SU(N) mode, where the spin dipole magnitude is allowed to vary.